+++ crates/core/src/use_cases.rs
@@ -1,4 +1,0 @@
-pub mod user;
-
-pub use user::UserUseCases;
-pub(crate) use user::UserUseCasesImpl;